added throwable eggs and refactor some projectile logic close #1473
